healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc

Former-commit-id: 7b3767ce593a49bee4429747bb9a197695fde667
Former-commit-id: 808f9c569dacd83b08ef55bb69717a029631f240
Former-commit-id: ca3b4373b65978d26624dc5629bac6233bf27eb1
Former-commit-id: ab3aba996b06e1de4ac22a061ec9b3755ec5739c
Former-commit-id: 3a8ed1cdd66afc54534c2dc3a13bd97df0ad3a98