init: Weaken property override security for the init  extension
Sometimes we need to override ro.* properties by using
our vendor init extension.

Previously there was a security check which was blocking
that.
To resolve the issue, we need to weaken the security check
during the execution of our vendor init extension.

This is safe because the vendor init extension gets
executed as part of init construction and it is
considered a trusted system component.