init: Check for fastbootd before spoofing safetynet props
The real prop values must be retained in recovery/fastbootd in
order for fastbootd to allow/deny flashing correctly based on the
bootloader lock state. This is accomplished by checking androidboot keys
in the kernel cmdline and bootconfig (necessary on Pixel 6), and not
spoofing anything if the boot isn't a normal full-blown Android boot.

@jhenrique09 - Adapt to PE

Change-Id: I66d23fd91d82906b00d5eb020668f01ae83ec31f

Former-commit-id: 33d4578679733fb2d6fd0fd9b7baba8fd5f0be57
Former-commit-id: b929a87f6ee30e21f795442d52c6a9ece2822c66
Former-commit-id: 07e94e583fcf1cd209f59aa272fd669947d03fe3
Former-commit-id: 46f8899dc68601a85aac7d585f1c5bf4e3e9b476