Reapply "Migrate the blkio controller to the v2 cgroup hierarchy"
This reverts commit a6f147d003c1678ecb4d729073cacced08d0b0c8 [formerly c7a6fe684c7cea3fb9fba6dbb5cceeb698bfd318].
Repply the blkio controller migration because it was not responsible
for the test failures that led to the revert. See also the following bugs:
* https://b.corp.google.com/issues/260143932
  (v2/android-virtual-infra/test_mapping/presubmit-avd test failure)
* https://b.corp.google.com/issues/264620181
  (CtsInitTestCases.RebootTest#StopServicesSIGKILL failure)

The only change compared with the previous version is that the io
controller has been declared optional. This is necessary because some
devices have a kernel that does not support the io controller.

Bug: 213617178
Test: Cuttlefish and various phones
Change-Id: I490740e1c9ee4f7bb5bb7afba721a083f952c8f2
Signed-off-by: Bart Van Assche <bvanassche@google.com>

Former-commit-id: d71f9e5fbbbc1b4a0f910c2c7606915bfeae68a4